[PAM-2196] Bytt ut blå rund knapp med "Vis treff" og "Endre søk" knapper på mobil. Plasser disse på toppen i en sticky buttonbar. På pc vises bare en "Til toppen" lenke på bunn av siden
