+++ src/search/Search.less
@@ -79,13 +79,17 @@
   white-space: nowrap;
 }
 
-
 .Search__header {
   display: flex;
   align-items: center;
   width: 100%;
   background-color: @background-green;
-  margin-bottom: @site-header-margin-bottom;
+}
+
+@media (min-width: @screen-md-min) {
+  .Search__header {
+    margin-bottom: @site-header-margin-bottom;
+  }
 }
 
 .Search__header__title {
@@ -116,3 +120,15 @@
   border-top: 1px solid #d8d8d8;
   margin: 1.25rem -1rem;
 }
+
+.Search__main__tiltoppen {
+  text-align: center;
+  padding: 2rem 0;
+}
+
+@media (max-width: @screen-sm-max) {
+  .Search__main__tiltoppen {
+    display: none;
+  }
+}
+
