--- src/search/backToTopButton/BackToTop.less
@@ -1,97 +1,0 @@
-@import "../../variables";
-
-.VisTreffButton,
-.TilToppenButton{
-  position: fixed;
-  bottom: 3.125rem; // Unngå at den kommer helt nederst på skjermen, siden klikk her vil åpne menu bar på iphone.
-  right: 0.5rem;
-  z-index: 100;
-  color: #fff;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  text-align: center;
-  height: 4.5rem;
-  width: 4.5rem;
-  border-radius: 50%;
-  background: rgba(0, 103, 197, 0.95);
-  text-decoration: none;
-  box-shadow: 0 0 8px 0 @navGra60;
-  letter-spacing: -0.3px;
-
-  &:hover {
-    background: #0074df;
-  }
-
-  &:active {
-    background: #005B82;
-  }
-
-  &:focus {
-    outline: 0;
-    box-shadow: 0 0 0 3px @navOransjeLighten40;
-  }
-}
-
-.TilToppenButton__inner,
-.VisTreffButton__inner {
-  width: 100%;
-}
-
-.TilToppenButton__inner{
-  margin-top: -0.5rem;
-}
-
-.VisTreffButton__inner{
-  margin-top: 0.5rem;
-}
-
-@media (min-width: @screen-md-min) {
-  .TilToppenButton {
-    bottom: 3rem;
-    right: 3rem;
-    height: 5.5rem;
-    width: 5.5rem;
-  }
-
-  .TilToppenButton__inner{
-    margin-top: -0.75rem;
-  }
-
-  .VisTreffButton {
-    display: none;
-  }
-}
-
-.VisTreffButton__count {
-  line-height: 1;
-  display: block;
-  font-weight: bold;
-}
-
-.VisTreffButton__treff {
-  line-height: 1;
-  display: block;
-}
-
-.TilToppenButton__text--mobile {
-  display: block;
-  line-height: 1;
-  padding: 0 1rem;
-}
-
-.TilToppenButton__text--pc {
-  display: none;
-  line-height: 1;
-}
-
-@media (min-width: @screen-md-min) {
-
-  .TilToppenButton__text--mobile {
-    display: none;
-  }
-
-  .TilToppenButton__text--pc {
-    display: block;
-  }
-}
