+++ src/search/searchResults/SearchResults.less
@@ -5,8 +5,16 @@
   padding: 1rem 0 1rem 0;
 }
 
-.SearchResults__end {
-  min-height: 15rem;
+@media (max-width: @screen-sm-max) {
+  .SearchResults__end {
+    min-height: 15rem;
+  }
+}
+
+@media (min-width: @screen-md-min) {
+  .SearchResults__end {
+    min-height: 2rem;
+  }
 }
 
 .SearchResults__spinner {
