+++ src/search/showResultsButton/ShowResultsButton.less
@@ -1,0 +1,22 @@
+@import "../../variables";
+
+@media (min-width: @screen-md-min) {
+  .ShowResultsButton {
+    display: none;
+  }
+}
+
+.ShowResultsButton__flex {
+  border-bottom: 1px solid @border-gray;
+  padding: 0.75rem 0;
+  background: #F6F5F5;
+  height: 3.25rem;
+  margin-bottom: 1rem;
+  display: flex;
+  width: 100%;
+  align-items: center;
+}
+
+.ShowResultsButton__flex__count {
+  flex: 1;
+}
